Logic changed and tests modified.
