+++ owncloudApp/src/main/java/com/owncloud/android/presentation/ui/security/PassCodeActivity.kt
@@ -503,11 +503,9 @@ class PassCodeActivity : AppCompatActivity(), IEnableBiometrics {
         when (optionSelected) {
             BiometricStatus.ENABLED_BY_USER -> {
                 resultIntent.putExtra(BIOMETRIC_ENABLED_FROM_DIALOG_EXTRA, true)
-                passCodeViewModel.setBiometricsState(enabled = true)
             }
             BiometricStatus.DISABLED_BY_USER -> {
                 resultIntent.putExtra(BIOMETRIC_ENABLED_FROM_DIALOG_EXTRA, false)
-                passCodeViewModel.setBiometricsState(enabled = false)
             }
         }
         finish()
